requirements.txt
apache-airflow==2.10.4
clickhouse-driver==0.2.9
pandas==2.1.4
numpy==1.26.4
setuptools==70.0.0
wheel==0.43.0